remove unused nav files, Add back icon, Add hover effect, Add active effect
